\name{indo.european.languages}
\alias{indo.european.languages}
\docType{data}
\title{ Hartigan (1975) Indo-European Languages }
\description{
  The table contains foreign language equivalent of the names associated
  with the column names.  This is Table 13.8 in Chapter 13 of Hartigan (1975)
  on page 243.
}
\usage{data(indo.european.languages)}
\format{
  A data frame with 13 observations on the following 17 variables.
  \describe{
    \item{\code{language}}{a character vector for the foreign language}
    \item{\code{all}}{a character vector for the foreign language equivalent}
    \item{\code{bad}}{a character vector for the foreign language equivalent}
    \item{\code{belly}}{a character vector for the foreign language equivalent}
    \item{\code{black}}{a character vector for the foreign language equivalent}
    \item{\code{bone}}{a character vector for the foreign language equivalent}
    \item{\code{day}}{a character vector for the foreign language equivalent}
    \item{\code{die}}{a character vector for the foreign language equivalent}
    \item{\code{drink}}{a character vector for the foreign language equivalent}
    \item{\code{ear}}{a character vector for the foreign language equivalent}
    \item{\code{eat}}{a character vector for the foreign language equivalent}
    \item{\code{egg}}{a character vector for the foreign language equivalent}
    \item{\code{eye}}{a character vector for the foreign language equivalent}
    \item{\code{father}}{a character vector for the foreign language equivalent}
    \item{\code{fish}}{a character vector for the foreign language equivalent}
    \item{\code{five}}{a character vector for the foreign language equivalent}
    \item{\code{foot}}{a character vector for the foreign language equivalent}
  }
}
\details{
  Hartigan suggest that the minimum mutation algorithm is applied to this data set.
}
\source{
  Hartigan, J. A. (1975). \emph{Clustering Algorithms}, John Wiley, New York.
  
  SPAETH2 Cluster Analysis Datasets \url{http://people.sc.fsu.edu/~jburkardt/datasets/spaeth2/spaeth2.html}
}
\references{
  Hartigan, J. A. (1975). \emph{Clustering Algorithms}, John Wiley, New York.
}
\examples{
data(indo.european.languages)
}
\keyword{datasets}
