\name{olympic.track.1896.1964}
\alias{olympic.track.1896.1964}
\docType{data}
\title{ Hartigan (1975) Olympic Track 1896 to 1964 }
\description{
  Olympic track times, in tenths of a second, were recorded
  orver the years.  This is Table 6.1 in Chapter 6 of Hartigan (1975)
  on page 131.
}
\usage{data(olympic.track.1896.1964)}
\format{
  A data frame with 16 observations on the following 8 variables.
  \describe{
    \item{\code{year}}{a character vector for the year}
    \item{\code{t.100m}}{a numeric vector for the winning time in the 100 m}
    \item{\code{t.200m}}{a numeric vector for the winning time in the 200 m}
    \item{\code{t.400m}}{a numeric vector for the winning time in the 400 m}
    \item{\code{t.800m}}{a numeric vector for the winning time in the 800 m}
    \item{\code{t.1500m}}{a numeric vector for the winning time in the 1500 m}
    \item{\code{t.5000m}}{a numeric vector for the winning time in the 5000 m}
    \item{\code{t.10000m}}{a numeric vector for the winning time in the 10000 m}
  }
}
\details{
  None.
}
\source{
  \emph{The World Almanac} (1966), New York World-Telegram, New York,
  
  SPAETH2 Cluster Analysis Datasets \url{http://people.sc.fsu.edu/~jburkardt/datasets/spaeth2/spaeth2.html}
}
\references{
  Hartigan, J. A. (1975). \emph{Clustering Algorithms}, John Wiley, New York.
}
\examples{
data(olympic.track.1896.1964)
}
\keyword{datasets}
